server/requirements.txt
django==4.0.4
django-cors-headers==3.11.0
django-axes==5.32.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-extensions==0.7.1
dj-rest-auth[with_social]==2.2.4

psycopg[binary]==3.0.12
psycopg==3.0.12
psycopg2-binary==2.9.3
uwsgi==2.0.20
botocore==1.24.46
boto3==1.21.46
Jinja2==3.1.1
validators==0.19.0
requests==2.27.1
celery==5.2.3
redis==4.3.1
simple-salesforce==1.11.6
cryptography==37.0.2
chardet==4.0.0